requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
ecpy
jinja2
SQLAlchemy>=1.4
